+++ LookupAnything/release-notes.md
@@ -9,6 +9,8 @@
   * collections menu.
 * Added times cooked/crafted to item lookups. (Thanks to watson81!)
 * Added 'needed for' support for Gourmet Chef and Craft Master achivements. (Thanks to watson81!)
+* After clicking a link in a lookup menu, closing the new lookup now returns you to the previous one.
+* Fixed previous menu not restored when `HideOnKeyUp` option is enabled.
 * Fixed debug fields showing wrong values in rare cases when an item was customised after it was spawned.
 * Improved translations. Thanks to Nanogamer7 (German)!